+++ temporalcli/commands.workflow.go
@@ -1,9 +1,7 @@
 package temporalcli
 
 import (
-	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"os/user"
 
@@ -16,19 +14,6 @@ import (
 	"go.temporal.io/api/query/v1"
 	"go.temporal.io/api/workflowservice/v1"
 	"go.temporal.io/sdk/client"
-)
-
-var (
-	resetTypesMap = map[string]interface{}{
-		"FirstWorkflowTask":  "",
-		"LastWorkflowTask":   "",
-		"LastContinuedAsNew": "",
-	}
-	resetReapplyTypesMap = map[string]interface{}{
-		"":       enums.RESET_REAPPLY_TYPE_SIGNAL, // default value
-		"Signal": enums.RESET_REAPPLY_TYPE_SIGNAL,
-		"None":   enums.RESET_REAPPLY_TYPE_NONE,
-	}
 )
 
 func (c *TemporalWorkflowCancelCommand) run(cctx *CommandContext, args []string) error {
@@ -123,59 +108,6 @@ func (c *TemporalWorkflowQueryCommand) run(cctx *CommandContext, args []string) 
 		return fmt.Errorf("failed to marshal query result: %w", err)
 	}
 	return cctx.Printer.PrintStructured(output, printer.StructuredOptions{})
-}
-
-func (c *TemporalWorkflowResetCommand) run(cctx *CommandContext, _ []string) error {
-	if c.Type.Value == "" && c.EventId <= 0 {
-		return errors.New("must specify either valid event id or reset type")
-	}
-	cl, err := c.Parent.ClientOptions.dialClient(cctx)
-	if err != nil {
-		return err
-	}
-	defer cl.Close()
-
-	resetBaseRunID := c.RunId
-	eventID := int64(c.EventId)
-	if c.Type.Value != "" {
-		resetBaseRunID, eventID, err = getResetEventIDByType(cctx, c.Type.Value, c.Parent.Namespace, c.WorkflowId, c.RunId, cl.WorkflowService())
-		if err != nil {
-			return fmt.Errorf("getting reset event ID by type failed: %w", err)
-		}
-	}
-	reapplyType := enums.RESET_REAPPLY_TYPE_SIGNAL
-	if c.ReapplyType.Value != "All" {
-		reapplyType, err = enums.ResetReapplyTypeFromString(c.ReapplyType.Value)
-		if err != nil {
-			return err
-		}
-	}
-
-	if eventID > 0 {
-		cctx.Printer.Printlnf("Resetting workflow %s to event ID %d", c.WorkflowId, eventID)
-	}
-
-	resp, err := cl.ResetWorkflowExecution(cctx, &workflowservice.ResetWorkflowExecutionRequest{
-		Namespace: c.Parent.Namespace,
-		WorkflowExecution: &common.WorkflowExecution{
-			WorkflowId: c.WorkflowId,
-			RunId:      resetBaseRunID,
-		},
-		Reason:                    fmt.Sprintf("%s: %s", username(), c.Reason),
-		WorkflowTaskFinishEventId: eventID,
-		RequestId:                 uuid.NewString(),
-		ResetReapplyType:          reapplyType,
-	})
-	if err != nil {
-		return fmt.Errorf("failed to reset workflow: %w", err)
-	}
-
-	if cctx.JSONOutput {
-		return cctx.Printer.PrintStructured(
-			resp,
-			printer.StructuredOptions{})
-	}
-	return nil
 }
 
 func (*TemporalWorkflowResetBatchCommand) run(*CommandContext, []string) error {
@@ -376,136 +308,4 @@ func startBatchJob(cctx *CommandContext, cl client.Client, req *workflowservice.
 	}
 	cctx.Printer.Printlnf("Started batch for job ID: %v", req.JobId)
 	return nil
-}
-
-func getResetEventIDByType(ctx context.Context, resetType, namespace, wid, rid string, wfsvc workflowservice.WorkflowServiceClient) (string, int64, error) {
-	switch resetType {
-	case "LastWorkflowTask":
-		return getLastWorkflowTaskEventID(ctx, namespace, wid, rid, wfsvc)
-	case "LastContinuedAsNew":
-		return getLastContinueAsNewID(ctx, namespace, wid, rid, wfsvc)
-	case "FirstWorkflowTask":
-		return getFirstWorkflowTaskEventID(ctx, namespace, wid, rid, wfsvc)
-	default:
-		return "", -1, fmt.Errorf("invalid reset type: %s", resetType)
-	}
-}
-
-// Returns event id of the last completed task or id of the next event after scheduled task.
-func getLastWorkflowTaskEventID(ctx context.Context, namespace, wid, rid string, wfsvc workflowservice.WorkflowServiceClient) (resetBaseRunID string, workflowTaskEventID int64, err error) {
-	resetBaseRunID = rid
-	req := workflowservice.GetWorkflowExecutionHistoryReverseRequest{
-		Namespace: namespace,
-		Execution: &common.WorkflowExecution{
-			WorkflowId: wid,
-			RunId:      rid,
-		},
-		MaximumPageSize: 250,
-		NextPageToken:   nil,
-	}
-
-	for more := true; more; more = len(req.NextPageToken) != 0 {
-		resp, err := wfsvc.GetWorkflowExecutionHistoryReverse(ctx, &req)
-		if err != nil {
-			return "", 0, fmt.Errorf("failed to get workflow execution history: %w", err)
-		}
-		for _, e := range resp.GetHistory().GetEvents() {
-			if e.GetEventType() == enums.EVENT_TYPE_WORKFLOW_TASK_COMPLETED {
-				workflowTaskEventID = e.GetEventId()
-				break
-			} else if e.GetEventType() == enums.EVENT_TYPE_WORKFLOW_TASK_SCHEDULED {
-				// if there is no task completed event, set it to first scheduled event + 1
-				workflowTaskEventID = e.GetEventId() + 1
-			}
-		}
-		req.NextPageToken = resp.NextPageToken
-	}
-	if workflowTaskEventID == 0 {
-		return "", 0, errors.New("unable to find any scheduled or completed task")
-	}
-	return
-}
-
-// Returns id of the first workflow task completed event or if it doesn't exist then id of the event after task scheduled event.
-func getFirstWorkflowTaskEventID(ctx context.Context, namespace, wid, rid string, wfsvc workflowservice.WorkflowServiceClient) (resetBaseRunID string, workflowTaskEventID int64, err error) {
-	resetBaseRunID = rid
-	req := workflowservice.GetWorkflowExecutionHistoryRequest{
-		Namespace: namespace,
-		Execution: &common.WorkflowExecution{
-			WorkflowId: wid,
-			RunId:      rid,
-		},
-		MaximumPageSize: 250,
-		NextPageToken:   nil,
-	}
-	for more := true; more; more = len(req.NextPageToken) != 0 {
-		resp, err := wfsvc.GetWorkflowExecutionHistory(ctx, &req)
-		if err != nil {
-			return "", 0, fmt.Errorf("failed to get workflow execution history: %w", err)
-		}
-		for _, e := range resp.GetHistory().GetEvents() {
-			if e.GetEventType() == enums.EVENT_TYPE_WORKFLOW_TASK_COMPLETED {
-				workflowTaskEventID = e.GetEventId()
-				return resetBaseRunID, workflowTaskEventID, nil
-			}
-			if e.GetEventType() == enums.EVENT_TYPE_WORKFLOW_TASK_SCHEDULED {
-				if workflowTaskEventID == 0 {
-					workflowTaskEventID = e.GetEventId() + 1
-				}
-			}
-		}
-		req.NextPageToken = resp.NextPageToken
-	}
-	if workflowTaskEventID == 0 {
-		return "", 0, errors.New("unable to find any scheduled or completed task")
-	}
-	return
-}
-
-func getLastContinueAsNewID(ctx context.Context, namespace, wid, rid string, wfsvc workflowservice.WorkflowServiceClient) (resetBaseRunID string, workflowTaskCompletedID int64, err error) {
-	// get first event
-	req := &workflowservice.GetWorkflowExecutionHistoryRequest{
-		Namespace: namespace,
-		Execution: &common.WorkflowExecution{
-			WorkflowId: wid,
-			RunId:      rid,
-		},
-		MaximumPageSize: 1,
-		NextPageToken:   nil,
-	}
-	resp, err := wfsvc.GetWorkflowExecutionHistory(ctx, req)
-	if err != nil {
-		return "", 0, fmt.Errorf("failed to get workflow execution history: %w", err)
-	}
-	firstEvent := resp.History.Events[0]
-	resetBaseRunID = firstEvent.GetWorkflowExecutionStartedEventAttributes().GetContinuedExecutionRunId()
-	if resetBaseRunID == "" {
-		return "", 0, errors.New("cannot use LastContinuedAsNew for workflow; workflow was not continued from another")
-	}
-
-	req = &workflowservice.GetWorkflowExecutionHistoryRequest{
-		Namespace: namespace,
-		Execution: &common.WorkflowExecution{
-			WorkflowId: wid,
-			RunId:      resetBaseRunID,
-		},
-		MaximumPageSize: 250,
-		NextPageToken:   nil,
-	}
-	for more := true; more; more = len(req.NextPageToken) != 0 {
-		resp, err := wfsvc.GetWorkflowExecutionHistory(ctx, req)
-		if err != nil {
-			return "", 0, fmt.Errorf("failed to get workflow execution history of previous execution (run id %s): %w", resetBaseRunID, err)
-		}
-		for _, e := range resp.GetHistory().GetEvents() {
-			if e.GetEventType() == enums.EVENT_TYPE_WORKFLOW_TASK_COMPLETED {
-				workflowTaskCompletedID = e.GetEventId()
-			}
-		}
-		req.NextPageToken = resp.NextPageToken
-	}
-	if workflowTaskCompletedID == 0 {
-		return "", 0, errors.New("unable to find WorkflowTaskCompleted event for previous execution")
-	}
-	return
 }
